// src/schemas/mst_mandatory_hdmf_Schemas.ts
export interface mst_mandatory_hdmfRequestBody{
    amount_start: number
    amount_end: number
    employee_contribution_percentage:number
    employer_contribution_percentage: number
    employee_contribution_value: number
    employer_contribution_value:number
    remarks: string
}

export interface mst_mandatory_hdmfID {
    id: number;
}